refactor: add filter module
